# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: C100 System API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /historian/archive:
    get:
      operationId: listArchiveEntries
      tags: [Historian]
      summary: List archive entries
      responses:
        "200":
          description: Archive entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchiveEntry"
    post:
      operationId: createArchiveEntry
      tags: [Historian]
      summary: Create an archive entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveEntryCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveEntry"
  /historian/archive/{id}:
    patch:
      operationId: patchArchiveEntry
      tags: [Historian]
      summary: Update an archive entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveEntryCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveEntry"
  /historian/archive/{id}/archive:
    post:
      operationId: archiveHistorianEntry
      tags: [Historian]
      summary: Soft-archive an archive entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveEntry"
  /historian/archive/{id}/restore:
    post:
      operationId: restoreHistorianEntry
      tags: [Historian]
      summary: Restore a soft-archived entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveEntry"
...
  /procedure/motions:
    get:
      operationId: listMotions
      tags: [Procedure]
      summary: List parliamentary motions
      responses:
        "200":
          description: Motions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Motion"
    post:
      operationId: createMotion
      tags: [Procedure]
      summary: Record a motion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motion"
  /procedure/motions/{id}:
    patch:
      operationId: patchMotion
      tags: [Procedure]
      summary: Update a motion
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotionCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motion"
  /procedure/rulings:
    get:
      operationId: listRulings
      tags: [Procedure]
      summary: List parliamentary rulings
      responses:
        "200":
          description: Rulings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParliamentaryRuling"
    post:
      operationId: createRuling
      tags: [Procedure]
      summary: Record a parliamentary ruling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParliamentaryRulingCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParliamentaryRuling"
  /procedure/rulings/{id}:
    patch:
      operationId: patchRuling
      tags: [Procedure]
      summary: Update a parliamentary ruling
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParliamentaryRulingCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParliamentaryRuling"
  /procedure/quorum:
    get:
      operationId: listQuorumRecords
      tags: [Procedure]
      summary: List quorum records
      responses:
        "200":
          description: Quorum records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuorumRecord"
    post:
      operationId: createQuorumRecord
      tags: [Procedure]
      summary: Record quorum check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuorumRecordCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuorumRecord"
  /procedure/quorum/{id}:
    get:
      operationId: getQuorumRecord
      tags: [Procedure]
      summary: Get a quorum record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Quorum record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuorumRecord"
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
...
  schemas:
...
    ArchiveEntry:
      type: object
      required: [id, title, description, eventDate, category, visibility, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
        category:
          type: string
        peopleText:
          type: [string, 'null']
        memberRefs:
          type: array
          items:
            type: integer
        storageKey:
          type: [string, 'null']
        originalFilename:
          type: [string, 'null']
        visibility:
          type: string
        tags:
          type: array
          items:
            type: string
        archivedAt:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArchiveEntryCreate:
      type: object
      required: [title, description, eventDate, category, visibility]
      properties:
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
        category:
          type: string
        peopleText:
          type: [string, 'null']
        storageKey:
          type: [string, 'null']
        originalFilename:
          type: [string, 'null']
        visibility:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Conduct schemas ──────────────────────────────────────────────────────
...
    Motion:
      type: object
      required: [id, motionText, result, voteYes, voteNo, voteAbstain, createdAt]
      properties:
        id:
          type: integer
        motionText:
          type: string
        result:
          type: string
        voteYes:
          type: integer
        voteNo:
          type: integer
        voteAbstain:
          type: integer
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    MotionCreate:
      type: object
      required: [motionText, result]
      properties:
        motionText:
          type: string
        result:
          type: string
        voteYes:
          type: integer
        voteNo:
          type: integer
        voteAbstain:
          type: integer
        notes:
          type: [string, 'null']
    ParliamentaryRuling:
      type: object
      required: [id, rulingText, authoritySource, createdAt]
      properties:
        id:
          type: integer
        rulingText:
          type: string
        authoritySource:
          type: string
        governanceRef:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    ParliamentaryRulingCreate:
      type: object
      required: [rulingText, authoritySource]
      properties:
        rulingText:
          type: string
        authoritySource:
          type: string
        governanceRef:
          type: [string, 'null']
    QuorumRecordCreate:
      type: object
      required: [totalMembership, quorumThreshold, membersPresent]
      properties:
        totalMembership:
          type: integer
        quorumThreshold:
          type: integer
        membersPresent:
          type: integer
        notes:
          type: [string, 'null']
    QuorumRecord:
      type: object
      required: [id, totalMembership, quorumThreshold, membersPresent, quorumMet, recordedAt]
      properties:
        id:
          type: integer
        totalMembership:
          type: integer
        quorumThreshold:
          type: integer
        membersPresent:
          type: integer
        quorumMet:
          type: integer
        notes:
          type: [string, 'null']
        recordedAt:
          type: string
          format: date-time
...