# lib/api-zod/src/generated/types/archiveEntry.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * C100 System API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArchiveEntry {
  id: number;
  title: string;
  description: string;
  eventDate: string;
  category: string;
  /** @nullable */
  peopleText?: string | null;
  memberRefs?: number[];
  /** @nullable */
  storageKey?: string | null;
  /** @nullable */
  originalFilename?: string | null;
  visibility: string;
  tags?: string[];
  /** @nullable */
  archivedAt?: string | null;
  createdAt: Date;
  updatedAt: Date;
}