# lib/api-zod/src/generated/types/parliamentaryRulingCreate.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * C100 System API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParliamentaryRulingCreate {
  rulingText: string;
  authoritySource: string;
  /** @nullable */
  governanceRef?: string | null;
}